requirements.txt
requests
openai
selenium
python-telegram-bot<=13
asyncio
sqlalchemy